+++ src/frameworks/next/index.ts
@@ -100,7 +100,7 @@ export async function build(
   });
 
   const reasonsForBackend = [];
-  const { distDir } = await getConfig(dir);
+  const { distDir, trailingSlash } = await getConfig(dir);
 
   if (await isUsingMiddleware(join(dir, distDir), false)) {
     reasonsForBackend.push("middleware");
@@ -199,7 +199,9 @@ export async function build(
     };
   });
 
-  const isEveryRedirectSupported = nextJsRedirects.every(isRedirectSupportedByHosting);
+  const isEveryRedirectSupported = nextJsRedirects
+    .filter((it) => !it.internal)
+    .every(isRedirectSupportedByHosting);
   if (!isEveryRedirectSupported) {
     reasonsForBackend.push("advanced redirects");
   }
@@ -267,7 +269,7 @@ export async function build(
     console.log("");
   }
 
-  return { wantsBackend, headers, redirects, rewrites, i18n: nextjsI18n };
+  return { wantsBackend, headers, redirects, rewrites, trailingSlash, i18n: nextjsI18n };
 }
 
 /**
@@ -482,5 +484,10 @@ async function getConfig(dir: string): Promise<NextConfig & { distDir: string }>
       }
     }
   }
-  return { distDir: ".next", ...config };
+  return {
+    distDir: ".next",
+    // trailingSlash defaults to false in Next.js: https://nextjs.org/docs/api-reference/next.config.js/trailing-slash
+    trailingSlash: false,
+    ...config,
+  };
 }