+++ CHANGELOG.md
@@ -1,2 +1,5 @@
-- Refactor Functions Emulator. (#5422)
-- Fix race condition when discovering functions. (#5444)
+- Refactors Functions Emulator. (#5422)
+- Fixes race condition when discovering functions. (#5444)
+- Fixes issue where `init firestore` was unecessarilly checking for default resource location. (#5230 and #5452)
+- Pass `trailingSlash` from Next.js config to `firebase.json` (#5445)
+- Don't use Next.js internal redirects for the backend test (#5445)